requirements.txt
Flask==2.2.3
google-api-python-client==2.88.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapiclient==1.13.4
watchdog==3.0.0
